refactor: changed form structure
