+++ src/apiCalls/apiURL.js
@@ -1,0 +1,1 @@
+export const API_URL = "https://sm-spring-api.herokuapp.com";
