--- src/redux/actions/companiesActions.jsx
@@ -1,2 +1,0 @@
-export const GET_COMPANIES = "GET_COMPANIES";
-export const companiesAction = (payload) => ({ type: GET_COMPANIES, payload });
